fix(crypto): widen Hash.pbkdf2 iteration counter to u64 (UA-1.3)
pbkdf2_sha256 took iterations as u32, and the native cast the caller's int
with 'as u32' — silently truncating any count above u32::MAX (e.g. 5_000_000_000
iterations ran as 705_032_704). PHP's hash_pbkdf2 takes the count as a native
int (i64) and uses it in full, so the truncation was also a byte-identity break
on the PHP leg for large counts. Widened the signature + cast to u64; the RFC
8018 block counter stays u32. No behavior change for counts <= u32::MAX; restores
run == runvm == php parity for larger counts. 19/19 crypto KATs green; crypto-mac
example 3-way byte-identical under php-8.5.8.

+++ src/native/hash.rs
@@ -440,8 +440,10 @@ fn hkdf_sha256(ikm: &[u8], salt: &[u8], info: &[u8], length: usize) -> Result<Ve
     Ok(okm)
 }
 
-/// PBKDF2-HMAC-SHA256 (RFC 8018 §5.2).
-fn pbkdf2_sha256(password: &[u8], salt: &[u8], iterations: u32, length: usize) -> Vec<u8> {
+/// PBKDF2-HMAC-SHA256 (RFC 8018 §5.2). `iterations` is `u64` (not `u32`): PHP's `hash_pbkdf2`
+/// takes the count as a native `int` (i64), so a `u32` cap would silently truncate a large
+/// iteration count and diverge from the PHP leg (UA-1.3). The RFC block counter stays `u32`.
+fn pbkdf2_sha256(password: &[u8], salt: &[u8], iterations: u64, length: usize) -> Vec<u8> {
     let mut out = Vec::with_capacity(length);
     let mut block_index: u32 = 1;
     while out.len() < length {
@@ -515,7 +517,7 @@ fn pbkdf2_native(a: &[Value], _: &mut String) -> Result<Value, String> {
     match a {
         [Value::Bytes(pw), Value::Bytes(salt), Value::Int(iters), len] if *iters > 0 => {
             let length = nonneg_len(len, "pbkdf2")?;
-            let dk = pbkdf2_sha256(pw, salt, *iters as u32, length);
+            let dk = pbkdf2_sha256(pw, salt, *iters as u64, length);
             Ok(Value::Bytes(std::rc::Rc::new(dk)))
         }
         _ => Err(
